Add GitHub Actions workflow for Docker image build and push
Create a new workflow in docker-build.yml to automate the building and pushing of Docker images on push events to the master branch. The workflow includes steps for checking out the code, setting version information, configuring QEMU and Docker Buildx, logging into Docker Hub, and building and pushing the Docker image with appropriate tags and build arguments.